refactor(dock): extract instance lifecycle
